Fixed put error

+++ view_operations.py
@@ -27,14 +27,14 @@ headers = {"Content-Type": "application/json"}
 def broadcast_put():
     if request.method == 'PUT':
         replica_view = os.getenv('VIEW')
-        address_to_be_added = request.json['socket-address']
+        address_to_be_added = request.args['socket-address']
         updated_view = replica_view + "," + address_to_be_added
         os.environ['VIEW'] = updated_view
         
         return {"message":"Replica added successfully to the view"}, 201
     elif request.method == 'DELETE':
         replica_view = os.getenv('VIEW')
-        addres_to_be_deleted = request.json['socket-address']
+        addres_to_be_deleted = request.args['socket-address']
         replica_view_list = replica_view.split(',')
         replica_view_list.remove(addres_to_be_deleted)
         hold_new_view = ",".join(replica_view_list)
@@ -54,7 +54,7 @@ def view_operations():
         return {"message":"View retrieved succesfully","view":replica_view}, 200
 
     elif request.method == 'PUT':
-        address_to_be_added = request.json['socket-address']
+        address_to_be_added = request.args['socket-address']
 
         if address_to_be_added in view_list:
             return {"error":"Socket address already exists in the view","message":"Error in PUT"}, 404
@@ -66,11 +66,11 @@ def view_operations():
             replica_view = os.getenv('VIEW')
 
             #Calling broadcasting function on a PUT operation
-            sending_broadcast_message(view_list,socket_address, address_to_be_added, replica_view)
+            sending_broadcast_message(view_list,socket_address, address_to_be_added, replica_view,"put")
             return {"message":"Replica added successfully to the view"}, 201
             
     elif request.method == 'DELETE':
-        address_to_delete = request.json['socket-address']
+        address_to_delete = request.args['socket-address']
 
         if address_to_delete not in view_list:
             return {"error":"Socket address does not exist in the view","message":"Error in DELETE"}, 404
@@ -83,7 +83,7 @@ def view_operations():
             replica_view = os.getenv('VIEW')
 
             #Calling broadcasting function 
-            sending_broadcast_message(view_list,socket_address, address_to_delete, replica_view)
+            sending_broadcast_message(view_list,socket_address, address_to_delete, replica_view,"delete")
             return {"message":"Replica deleted successfully from the view"}, 200
 
 #returns replicas view and url
@@ -95,14 +95,17 @@ def get_view(ip_address):
     return response_view, address_to_search
 
 #Loop to broadcast a message 
-def sending_broadcast_message(view_list,socket_address, address_to_delete, replica_view):
+def sending_broadcast_message(view_list,socket_address, address_to_delete, replica_view, type_request):
     for ip in view_list:
         if ip != socket_address and ip != address_to_delete:
             response_view, address = get_view(ip)
 
             if response_view != replica_view:
                 address_to_search = address + "/broadcast?socket-address=" + address_to_delete
-                requests.delete(address_to_search, headers=headers, json=request.json)
+                if type_request == 'delete':
+                    requests.delete(address_to_search, headers=headers, json=request.json)
+                else:
+                    requests.put(address_to_search, headers=headers, json=request.json)
 
 
 if __name__=='__main__':
